requirements.txt
numpy==1.18.5
matplotlib==3.3.3
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cpu
tensorboard==2.5.0
seaborn==0.11.1
pyyaml